Кобминирует редьюсеры
