+++ src/index.js
@@ -4,11 +4,12 @@ import App from "./components/app/app.jsx";
 import {mock, availableGenres} from "./mocks/films.js";
 import {createStore, applyMiddleware} from "redux";
 import {Provider} from "react-redux";
-import {reducer} from "./reducer/reducer.js";
+import reducer from "./reducer/reducer.js";
 import withSelectedFilm from "../src/hoc/with-selected-film/with-selected-film.js";
 import {createAPI} from "./api.js";
 import thunk from "redux-thunk";
-import {ActionCreator, AuthorizationStatus, Operation} from "./reducer/reducer.js";
+import {Operation as DataOperation} from "./reducer/data/data.js";
+import {Operation as UserOperation, ActionCreator, AuthorizationStatus} from "./reducer/user/user.js";
 
 const onUnauthorized = () => {
   store.dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.NO_AUTH));
@@ -22,9 +23,9 @@ const store = createStore(
 );
 
 
-store.dispatch(Operation.loadFilms());
-store.dispatch(Operation.checkAuth());
-store.dispatch(Operation.loadPromoFilm());
+store.dispatch(DataOperation.loadFilms());
+store.dispatch(UserOperation.checkAuth());
+store.dispatch(DataOperation.loadPromoFilm());
 
 
 const WrappedApp = withSelectedFilm(App);
