+++ src/reducer/user/user.js
@@ -1,0 +1,64 @@
+import {extend} from "../../utils.js";
+
+const AuthorizationStatus = {
+  AUTH: `AUTH`,
+  NO_AUTH: `NO_AUTH`,
+};
+
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.NO_AUTH,
+};
+
+const ActionType = {
+  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
+};
+
+
+const Operation = {
+  checkAuth: () => (dispatch, getState, api) => {
+    return api.get(`/login`)
+      .then(() => {
+        dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.AUTH));
+      })
+      .catch((err) => {
+        throw err;
+      });
+  },
+  login: (authData) => (dispatch, getState, api) => {
+    return api.post(`/login`, {
+      email: authData.login,
+      password: authData.password,
+    })
+      .then(() => {
+        dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.AUTH));
+      });
+  },
+};
+
+
+const ActionCreator = {
+  requireAuthorization: (status) => {
+    return {
+      type: ActionType.REQUIRED_AUTHORIZATION,
+      payload: status,
+    };
+  },
+};
+
+
+const reducer = (state = initialState, action) => {
+
+  switch (action.type) {
+    case ActionType.REQUIRED_AUTHORIZATION: {
+      return extend(state, {
+        authorizationStatus: action.payload
+      });
+    }
+  }
+
+  return state;
+};
+
+
+export {reducer, ActionCreator, ActionType, Operation, AuthorizationStatus};
